# web/openapi.json
{
  "components": {
    "schemas": {
...
      "CommonResponse_Any_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[Any]",
        "type": "object"
      },
...
      "CommonResponse_DeleteWorkProjectResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeleteWorkProjectResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[DeleteWorkProjectResponse]",
        "type": "object"
      },
...
      "CommonResponse_QueryWorkProjectsResponse_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/QueryWorkProjectsResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[QueryWorkProjectsResponse]",
        "type": "object"
      },
...
      "CommonResponse_WorkProjectSchema_": {
        "properties": {
          "code": {
            "default": 200,
            "title": "Code",
            "type": "integer"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkProjectSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "default": "success",
            "title": "Message",
            "type": "string"
          }
        },
        "title": "CommonResponse[WorkProjectSchema]",
        "type": "object"
      },
...
      "CreateWorkProjectRequest": {
        "properties": {
          "description": {
            "default": "",
            "maxLength": 2000,
            "title": "Description",
            "type": "string"
          },
          "name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/WorkProjectTypeSchema",
            "default": "penetration_test"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreateWorkProjectRequest",
        "type": "object"
      },
...
      "DeleteWorkProjectResponse": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "deleted"
        ],
        "title": "DeleteWorkProjectResponse",
        "type": "object"
      },
...
      "QueryWorkProjectsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkProjectSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "size",
          "items"
        ],
        "title": "QueryWorkProjectsResponse",
        "type": "object"
      },
...
      "WorkProjectSchema": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/WorkProjectStatusSchema"
          },
          "type": {
            "$ref": "#/components/schemas/WorkProjectTypeSchema"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "session_id",
          "description",
          "status",
          "type",
          "created_at",
          "updated_at"
        ],
        "title": "WorkProjectSchema",
        "type": "object"
      },
      "WorkProjectStatusSchema": {
        "enum": [
          "working",
          "completed",
          "failed",
          "canceled"
        ],
        "title": "WorkProjectStatusSchema",
        "type": "string"
      },
      "WorkProjectTypeSchema": {
        "enum": [
          "penetration_test",
          "source_code_audit"
        ],
        "title": "WorkProjectTypeSchema",
        "type": "string"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Z3r0 - AI + Security Attack Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/work-projects": {
      "get": {
        "description": "query work projects",
        "operationId": "query_work_projects_route_api_work_projects_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "keyword",
            "required": false,
            "schema": {
              "default": "",
              "title": "Keyword",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_QueryWorkProjectsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Query Work Projects Route",
        "tags": [
          "work-projects"
        ]
      },
      "post": {
        "description": "create work project",
        "operationId": "create_work_project_handler_api_work_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkProjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_WorkProjectSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create Work Project Handler",
        "tags": [
          "work-projects"
        ]
      }
    },
    "/api/work-projects/{id}": {
      "delete": {
        "description": "delete work project",
        "operationId": "delete_work_project_handler_api_work_projects__id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_DeleteWorkProjectResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete Work Project Handler",
        "tags": [
          "work-projects"
        ]
      }
    },
    "/api/work-projects/{id}/cancel": {
      "post": {
        "description": "cancel active work project",
        "operationId": "cancel_work_project_handler_api_work_projects__id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_WorkProjectSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Cancel Work Project Handler",
        "tags": [
          "work-projects"
        ]
      }
    },
    "/api/work-projects/{id}/retry": {
      "post": {
        "description": "retry failed work project",
        "operationId": "retry_work_project_handler_api_work_projects__id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "title": "Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_WorkProjectSchema_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Work project not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse_Any_"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Retry Work Project Handler",
        "tags": [
          "work-projects"
        ]
      }
    }
  }
}